+++ src/nveditor/nveditor_globals.py
@@ -4,8 +4,12 @@ Copyright (c) Peter Triesberger
 For further information see https://github.com/peter88213/nv_editor
 License: GNU GPLv3 (https://www.gnu.org/licenses/gpl-3.0.en.html)
 """
+from tkinter import font as tkFont
+
 from nveditor.nveditor_locale import _
+
+prefs = {}
 
 FEATURE = _('Section Editor')
 ICON = 'editor'
-
+DEFAULT_FONT = 'Courier'