+++ tools/build.py
@@ -13,7 +13,7 @@ import sys
 sys.path.insert(0, f'{os.getcwd()}/../../novelibre/tools')
 from package_builder import PackageBuilder
 
-VERSION = '5.10.1'
+VERSION = '5.10.2'
 
 
 class PluginBuilder(PackageBuilder):